rename: AppConfig 파일 config 패키지로 이동
